chore: Move standalone fns to corresp. impl blocks
