+++ zkevm-circuits/src/evm_circuit.rs
@@ -14,7 +14,7 @@ pub(crate) mod util;
 pub mod table;
 
 use crate::table::{BlockTable, BytecodeTable, CopyTable, ExpTable, KeccakTable, RwTable, TxTable};
-use crate::util::{log2_ceil, Challenges, SubCircuit, SubCircuitConfig};
+use crate::util::{Challenges, SubCircuit, SubCircuitConfig};
 pub use crate::witness;
 use bus_mapping::evm::OpcodeId;
 use eth_types::Field;
@@ -271,7 +271,6 @@ pub mod test {
     use super::*;
     use crate::{
         evm_circuit::{witness::Block, EvmCircuitConfig},
-        exp_circuit::OFFSET_INCREMENT,
         table::{BlockTable, BytecodeTable, CopyTable, ExpTable, KeccakTable, RwTable, TxTable},
         util::Challenges,
     };
@@ -385,90 +384,21 @@ pub mod test {
     }
 
     impl<F: Field> EvmCircuit<F> {
-        pub fn get_num_rows_required(block: &Block<F>) -> usize {
-            let mut cs = ConstraintSystem::default();
-            let config = EvmCircuit::<F>::configure(&mut cs);
-            config.0.get_num_rows_required(block)
+        pub fn get_test_cicuit_from_block(block: Block<F>) -> Self {
+            let fixed_table_tags = detect_fixed_table_tags(&block);
+
+            EvmCircuit::<F>::new_dev(block, fixed_table_tags)
         }
-
-        pub fn get_active_rows(block: &Block<F>) -> (Vec<usize>, Vec<usize>) {
-            let mut cs = ConstraintSystem::default();
-            let config = EvmCircuit::<F>::configure(&mut cs);
-            config.0.get_active_rows(block)
-        }
-    }
-
-    // TODO: Move this to `Block` under cfg[test]
-    pub fn get_test_degree<F: Field>(block: &Block<F>) -> u32 {
-        let num_rows_required_for_execution_steps: usize =
-            EvmCircuit::<F>::get_num_rows_required(block);
-        let num_rows_required_for_rw_table: usize = block.circuits_params.max_rws;
-        let num_rows_required_for_fixed_table: usize = detect_fixed_table_tags(block)
-            .iter()
-            .map(|tag| tag.build::<F>().count())
-            .sum();
-        let num_rows_required_for_bytecode_table: usize = block
-            .bytecodes
-            .values()
-            .map(|bytecode| bytecode.bytes.len() + 1)
-            .sum();
-        let num_rows_required_for_copy_table: usize =
-            block.copy_events.iter().map(|c| c.bytes.len() * 2).sum();
-        let num_rows_required_for_keccak_table: usize = block.keccak_inputs.len();
-        let num_rows_required_for_tx_table: usize =
-            block.txs.iter().map(|tx| 9 + tx.call_data.len()).sum();
-        let num_rows_required_for_exp_table: usize = block
-            .exp_events
-            .iter()
-            .map(|e| e.steps.len() * OFFSET_INCREMENT)
-            .sum();
-
-        const NUM_BLINDING_ROWS: usize = 64;
-
-        let rows_needed: usize = itertools::max([
-            num_rows_required_for_execution_steps,
-            num_rows_required_for_rw_table,
-            num_rows_required_for_fixed_table,
-            num_rows_required_for_bytecode_table,
-            num_rows_required_for_copy_table,
-            num_rows_required_for_keccak_table,
-            num_rows_required_for_tx_table,
-            num_rows_required_for_exp_table,
-        ])
-        .unwrap();
-
-        let k = log2_ceil(NUM_BLINDING_ROWS + rows_needed);
-        log::debug!(
-            "num_rows_requred_for rw_table={}, fixed_table={}, bytecode_table={}, \
-            copy_table={}, keccak_table={}, tx_table={}, exp_table={}",
-            num_rows_required_for_rw_table,
-            num_rows_required_for_fixed_table,
-            num_rows_required_for_bytecode_table,
-            num_rows_required_for_copy_table,
-            num_rows_required_for_keccak_table,
-            num_rows_required_for_tx_table,
-            num_rows_required_for_exp_table
-        );
-        log::debug!("evm circuit uses k = {}, rows = {}", k, rows_needed);
-        k
-    }
-
-    pub fn get_test_cicuit_from_block<F: Field>(block: Block<F>) -> EvmCircuit<F> {
-        let fixed_table_tags = detect_fixed_table_tags(&block);
-
-        EvmCircuit::<F>::new_dev(block, fixed_table_tags)
     }
 }
 
 #[cfg(test)]
 mod evm_circuit_stats {
-
-    use super::*;
     use crate::evm_circuit::step::ExecutionState;
     use crate::test_util::CircuitTestBuilder;
 
     use eth_types::{bytecode, evm_types::OpcodeId, geth_types::GethData};
-    use halo2_proofs::halo2curves::bn256::Fr;
+
     use mock::test_ctx::{helpers::*, TestContext};
     use strum::IntoEnumIterator;
