+++ zkevm-circuits/src/witness/block.rs
@@ -1,6 +1,9 @@
 use std::collections::HashMap;
 
-use crate::{evm_circuit::util::rlc, table::BlockContextFieldTag};
+use crate::{
+    evm_circuit::{detect_fixed_table_tags, evm_circuit::util::rlc, EvmCircuit},
+    table::BlockContextFieldTag,
+};
 use bus_mapping::{
     circuit_input_builder::{self, CircuitsParams, CopyEvent, ExpEvent},
     Error,
@@ -68,6 +71,72 @@ impl<F: Field> Block<F> {
         }
     }
 }
+
+#[cfg(feature = "test")]
+use crate::exp_circuit::OFFSET_INCREMENT;
+#[cfg(feature = "test")]
+use crate::util::log2_ceil;
+
+#[cfg(feature = "test")]
+impl<F: Field> Block<F> {
+    /// Obtains the expected Circuit degree needed in order to be able to test
+    /// the EvmCircuit with this block without needing to configure the
+    /// `ConstraintSystem`.
+    pub fn get_test_degree(&self) -> u32 {
+        let num_rows_required_for_execution_steps: usize =
+            EvmCircuit::<F>::get_num_rows_required(self);
+        let num_rows_required_for_rw_table: usize = self.circuits_params.max_rws;
+        let num_rows_required_for_fixed_table: usize = detect_fixed_table_tags(self)
+            .iter()
+            .map(|tag| tag.build::<F>().count())
+            .sum();
+        let num_rows_required_for_bytecode_table: usize = self
+            .bytecodes
+            .values()
+            .map(|bytecode| bytecode.bytes.len() + 1)
+            .sum();
+        let num_rows_required_for_copy_table: usize =
+            self.copy_events.iter().map(|c| c.bytes.len() * 2).sum();
+        let num_rows_required_for_keccak_table: usize = self.keccak_inputs.len();
+        let num_rows_required_for_tx_table: usize =
+            self.txs.iter().map(|tx| 9 + tx.call_data.len()).sum();
+        let num_rows_required_for_exp_table: usize = self
+            .exp_events
+            .iter()
+            .map(|e| e.steps.len() * OFFSET_INCREMENT)
+            .sum();
+
+        const NUM_BLINDING_ROWS: usize = 64;
+
+        let rows_needed: usize = itertools::max([
+            num_rows_required_for_execution_steps,
+            num_rows_required_for_rw_table,
+            num_rows_required_for_fixed_table,
+            num_rows_required_for_bytecode_table,
+            num_rows_required_for_copy_table,
+            num_rows_required_for_keccak_table,
+            num_rows_required_for_tx_table,
+            num_rows_required_for_exp_table,
+        ])
+        .unwrap();
+
+        let k = log2_ceil(NUM_BLINDING_ROWS + rows_needed);
+        log::debug!(
+            "num_rows_requred_for rw_table={}, fixed_table={}, bytecode_table={}, \
+            copy_table={}, keccak_table={}, tx_table={}, exp_table={}",
+            num_rows_required_for_rw_table,
+            num_rows_required_for_fixed_table,
+            num_rows_required_for_bytecode_table,
+            num_rows_required_for_copy_table,
+            num_rows_required_for_keccak_table,
+            num_rows_required_for_tx_table,
+            num_rows_required_for_exp_table
+        );
+        log::debug!("evm circuit uses k = {}, rows = {}", k, rows_needed);
+        k
+    }
+}
+
 /// Block context for execution
 #[derive(Debug, Default, Clone)]
 pub struct BlockContext {
